modified the place for all functions
