+++ thewalrus/tests/test_fock_gradients.py
@@ -283,174 +283,3 @@ def test_two_mode_squeezing_values(tol):
     T = two_mode_squeezing(r, theta, cutoff)
     expected = ((np.tanh(r) * np.exp(1j * theta)) ** np.arange(cutoff)) / np.cosh(r)
     assert np.allclose(np.diag(T[:, :, 0, 0]), expected, atol=tol, rtol=0)
-
-
-def test_gaussian_gate_values_with_single_mode_squeezing(tol):
-    """Tests the transforamtion matrix of gaussian gate. This test is particular for the single mode squeezing gate (S(zeta)) and other parameters are 0s here."""
-    cutoff = 5
-    zeta = 0.3 + 1j * 0.2
-    r = np.abs(zeta)
-    delta = np.angle(zeta)
-    expected = squeezing(r, delta, cutoff)
-    tanhr = np.tanh(r)
-    sechr = 1 / np.cosh(r)
-    C = np.sqrt(sechr)
-    mu = np.zeros(2).T
-    Sigma = np.array([[np.exp(1j * delta) * tanhr, -sechr], [-sechr, -np.exp(-1j * delta) * tanhr]])
-    T = gaussian_gate(C, mu, Sigma, cutoff)
-    assert np.allclose(T, expected, atol=tol, rtol=0)
-
-
-def test_gaussian_gate_values_with_single_mode_displacement(tol):
-    """Tests the transforamtion matrix of gaussian gate. This test is particular for the single mode displacement gate (D(gamma)) and other parameters are 0s here."""
-    cutoff = 4
-    gamma = 0.2 - 1j * 0.8
-    expected = displacement(np.abs(gamma), np.angle(gamma), cutoff)
-    C = np.exp(-0.5 * np.abs(gamma) ** 2)
-    mu = np.array([gamma, -np.conj(gamma)]).T
-    Sigma = np.array([[0, -1], [-1, 0]])
-    T = gaussian_gate(C, mu, Sigma, cutoff)
-    assert np.allclose(T, expected, atol=tol, rtol=0)
-
-
-def test_gaussian_gate_values_with_beamsplitter(tol):
-    """Tests the transforamtion matrix of gaussian gate. This test is particular for two-mode beamsplitter (BS(theta, phi)) and other parameters are 0s here."""
-    # Special case: BS
-    cutoff = 4
-    theta = np.pi / 4
-    phi = np.pi / 2
-    expected = beamsplitter(theta, phi, cutoff)
-    ct = np.cos(theta)
-    st = np.sin(theta) * np.exp(1j * phi)
-    V = np.array([[ct, -np.conj(st)], [st, ct]])
-    C = 1
-    mu = np.zeros(4).T
-    Sigma = -np.block([[np.zeros((2, 2)), V], [V.T, np.zeros((2, 2))]])
-    T = gaussian_gate(C, mu, Sigma, cutoff)
-    assert np.allclose(T, expected, atol=tol, rtol=0)
-
-
-def test_grad_gaussian_gate_with_single_mode_squeezing(tol):
-    """Tests the gradients of gaussian gate. This test is particular for the single mode squeezing gate (S(zeta)) and other parameters are 0s here. The gradients of parameters are tested by finite differences"""
-    # Special case: single-mode squeezing (zeta, gamma=0, phi=0)
-    cutoff = 6
-    zeta = 0.6 - 1j * 0.2
-    r = np.abs(zeta)
-    delta = np.angle(zeta)
-    T = squeezing(r, delta, cutoff)
-    tanhr = np.tanh(r)
-    sechr = 1 / np.cosh(r)
-    C = np.sqrt(sechr)
-    mu = np.zeros(2).T
-    Sigma = np.array([[np.exp(1j * delta) * tanhr, -sechr], [-sechr, -np.exp(-1j * delta) * tanhr]])
-    grad_C, grad_mu, grad_Sigma = grad_gaussian_gate(T, C, mu, Sigma, cutoff, dtype=np.complex128)
-    delta_plus = 0.00001 + 1j * 0.00001
-    expected_grad_C = (
-        gaussian_gate(C + delta_plus, mu, Sigma, cutoff)
-        - gaussian_gate(C - delta_plus, mu, Sigma, cutoff)
-    ) / (2 * delta_plus)
-    assert np.allclose(grad_C, expected_grad_C, atol=tol, rtol=0)
-    expected_grad_mu = (
-        gaussian_gate(C, mu + delta_plus, Sigma, cutoff)
-        - gaussian_gate(C, mu - delta_plus, Sigma, cutoff)
-    ) / (2 * delta_plus)
-    assert np.allclose(grad_mu, expected_grad_mu, atol=tol, rtol=0)
-    expected_grad_Sigma = (
-        gaussian_gate(C, mu, Sigma + delta_plus, cutoff)
-        - gaussian_gate(C, mu, Sigma - delta_plus, cutoff)
-    ) / (2 * delta_plus)
-    assert np.allclose(grad_Sigma, expected_grad_Sigma, atol=tol, rtol=0)
-
-
-def test_grad_gaussian_gate_with_beamsplitter(tol):
-    """Tests the gradients of gaussian gate. This test is particular for two-mode beamsplitter (BS(theta, phi)) and other parameters are 0s here. The gradients of parameters are tested by finite differences"""
-    # Special case: BS
-    cutoff = 5
-    theta = np.pi / 4
-    phi = np.pi / 2
-    T = beamsplitter(theta, phi, cutoff)
-    ct = np.cos(theta)
-    st = np.sin(theta) * np.exp(1j * phi)
-    V = np.array([[ct, -np.conj(st)], [st, ct]])
-    C = 1
-    mu = np.zeros(4).T
-    Sigma = -np.block([[np.zeros((2, 2)), V], [V.T, np.zeros((2, 2))]])
-    grad_C, grad_mu, grad_Sigma = grad_gaussian_gate(T, C, mu, Sigma, cutoff, dtype=np.complex128)
-    delta_plus = 0.00001 + 1j * 0.00001
-    expected_grad_C = (
-        gaussian_gate(C + delta_plus, mu, Sigma, cutoff)
-        - gaussian_gate(C - delta_plus, mu, Sigma, cutoff)
-    ) / (2 * delta_plus)
-    assert np.allclose(grad_C, expected_grad_C, atol=tol, rtol=0)
-    expected_grad_mu = (
-        gaussian_gate(C, mu + delta_plus, Sigma, cutoff)
-        - gaussian_gate(C, mu - delta_plus, Sigma, cutoff)
-    ) / (2 * delta_plus)
-    assert np.allclose(grad_mu, expected_grad_mu, atol=tol, rtol=0)
-    expected_grad_Sigma = (
-        gaussian_gate(C, mu, Sigma + delta_plus, cutoff)
-        - gaussian_gate(C, mu, Sigma - delta_plus, cutoff)
-    ) / (2 * delta_plus)
-    assert np.allclose(grad_Sigma, expected_grad_Sigma, atol=tol, rtol=0)
-
-
-def choi_trick(S, d):
-    """Function to help the test of gaussian gate with symplectic matrix, to get the parameter C, mu, Sigma of gaussian gate from S, d"""
-    num_mode = S.shape[0] // 2
-    choi_r = np.arcsinh(1.0)
-    ch = np.cosh(choi_r) * np.identity(num_mode)
-    sh = np.sinh(choi_r) * np.identity(num_mode)
-    zh = np.zeros([num_mode, num_mode])
-    Schoi = np.block([[ch, sh, zh, zh], [sh, ch, zh, zh], [zh, zh, ch, -sh], [zh, zh, -sh, ch]])
-    Sxx = S[:num_mode, :num_mode]
-    Sxp = S[:num_mode, num_mode:]
-    Spx = S[num_mode:, :num_mode]
-    Spp = S[num_mode:, num_mode:]
-    idl = np.identity(num_mode)
-    S_exp = (
-        np.block(
-            [
-                [Sxx, zh, Sxp, zh],
-                [zh, idl, zh, zh],
-                [Spx, zh, Spp, zh],
-                [zh, zh, zh, idl],
-            ]
-        )
-        @ Schoi
-    )
-    choi_cov = 0.5 * S_exp @ S_exp.T
-    idl = np.identity(2 * num_mode)
-    R = np.sqrt(0.5) * np.block([[idl, 1j * idl], [idl, -1j * idl]])
-    sigma = R @ choi_cov @ R.conj().T
-    zh = np.zeros([2 * num_mode, 2 * num_mode])
-    X = np.block([[zh, idl], [idl, zh]])
-    sigma_Q = sigma + 0.5 * np.identity(4 * num_mode)
-    A_mat = X @ (np.identity(4 * num_mode) - np.linalg.inv(sigma_Q))
-    E = np.diag(np.concatenate([np.ones([num_mode]), np.ones([num_mode]) / np.tanh(choi_r)]))
-    Sigma = -(E @ A_mat[: 2 * num_mode, : 2 * num_mode] @ E).conj()
-    mu = np.concatenate(
-        [Sigma[:num_mode, :num_mode] @ d.conj() + d.T, Sigma[num_mode:, :num_mode] @ d.conj()]
-    )
-    alpha = np.concatenate([d, np.zeros(num_mode)])
-    zeta = alpha + Sigma @ np.conj(alpha)
-    C = np.sqrt(
-        np.sqrt(
-            np.linalg.det(
-                np.eye(num_mode)
-                - Sigma[:num_mode, :num_mode] @ np.conj(Sigma[:num_mode, :num_mode])
-            )
-        )
-    ) * np.exp(-0.5 * np.sum(np.conj(alpha) * zeta))
-    return C, mu, Sigma
-
-
-def test_gaussian_gate_with_Symplectic_matrix(tol):
-    """Tests of the gaussian gate. This test is for arbitraty symplectic matrix and displacement vector as input and compare the gate with fock_tensor function"""
-    num_mode = 3
-    cutoff = 5
-    S = random_symplectic(num_mode)
-    d = np.random.random(num_mode) + 1j * np.random.random(num_mode)
-    _gaussian_gate = fock_tensor(S, d, cutoff)
-    C, mu, Sigma = choi_trick(S, d)
-    expected_gaussian_gate = gaussian_gate(C, mu, Sigma, cutoff)
-    assert np.allclose(_gaussian_gate, expected_gaussian_gate, atol=tol, rtol=0)
